+++ src/makefile.osx
@@ -110,7 +110,8 @@ OBJS= \
     obj/scrypt-arm.o \
     obj/scrypt-x86.o \
     obj/scrypt-x86_64.o \
-    obj/additionalfee.o
+    obj/additionalfee.o \
+    obj/messages.o
 
 ifndef USE_UPNP
 	override USE_UPNP = -
