+++ src/makefile.unix
@@ -153,7 +153,8 @@ OBJS= \
     obj/scrypt-arm.o \
     obj/scrypt-x86.o \
     obj/scrypt-x86_64.o \
-    obj/additionalfee.o
+    obj/additionalfee.o \
+    obj/messages.o
 
 all: FlyCoind
