+++ src/net.cpp
@@ -11,6 +11,7 @@
 #include "strlcpy.h"
 #include "addrman.h"
 #include "ui_interface.h"
+#include "messages.h"
 
 #ifdef WIN32
 #include <string.h>
